+++ packages/plugin-dev/src/common/plugin-dev-protocol.ts
@@ -15,6 +15,7 @@
  ********************************************************************************/
 
 import { JsonRpcServer } from '@theia/core/lib/common/messaging/proxy-factory';
+import { DebugPluginConfiguration } from '@theia/debug/lib/browser/debug-contribution';
 import { PluginMetadata } from '@theia/plugin-ext/lib/common/plugin-protocol';
 
 export const hostedServicePath = '/services/plugin-dev';
@@ -22,7 +23,7 @@ export const HostedPluginServer = Symbol('HostedPluginServer');
 export interface HostedPluginServer extends JsonRpcServer<HostedPluginClient> {
     getHostedPlugin(): Promise<PluginMetadata | undefined>;
     runHostedPluginInstance(uri: string): Promise<string>;
-    runDebugHostedPluginInstance(uri: string, debugConfig: DebugConfiguration): Promise<string>;
+    runDebugHostedPluginInstance(uri: string, debugConfig: DebugPluginConfiguration): Promise<string>;
     terminateHostedPluginInstance(): Promise<void>;
     isHostedPluginInstanceRunning(): Promise<boolean>;
     getHostedPluginInstanceURI(): Promise<string>;
@@ -36,9 +37,4 @@ export interface HostedPluginServer extends JsonRpcServer<HostedPluginClient> {
 }
 
 export interface HostedPluginClient {
-}
-
-export interface DebugConfiguration {
-    port?: number;
-    debugMode?: string;
 }