+++ packages/plugin-dev/src/node/hosted-instance-manager.ts
@@ -24,13 +24,15 @@ import * as request from 'request';
 import URI from '@theia/core/lib/common/uri';
 import { ContributionProvider } from '@theia/core/lib/common/contribution-provider';
 import { HostedPluginUriPostProcessor, HostedPluginUriPostProcessorSymbolName } from './hosted-plugin-uri-postprocessor';
-import { DebugConfiguration } from '../common';
 import { environment } from '@theia/core';
 import { FileUri } from '@theia/core/lib/node/file-uri';
 import { LogType } from '@theia/plugin-ext/lib/common/types';
 import { HostedPluginSupport } from '@theia/plugin-ext/lib/hosted/node/hosted-plugin';
 import { MetadataScanner } from '@theia/plugin-ext/lib/hosted/node/metadata-scanner';
+import { DebugPluginConfiguration } from '@theia/debug/lib/browser/debug-contribution';
 import { HostedPluginProcess } from '@theia/plugin-ext/lib/hosted/node/hosted-plugin-process';
+
+const DEFAULT_HOSTED_PLUGIN_PORT = 3030;
 
 export const HostedInstanceManager = Symbol('HostedInstanceManager');
 
@@ -58,7 +60,7 @@ export interface HostedInstanceManager {
      * @param debugConfig debug configuration
      * @returns uri where new Theia instance is run
      */
-    debug(pluginUri: URI, debugConfig: DebugConfiguration): Promise<URI>;
+    debug(pluginUri: URI, debugConfig: DebugPluginConfiguration): Promise<URI>;
 
     /**
      * Terminates hosted plugin instance.
@@ -118,11 +120,11 @@ export abstract class AbstractHostedInstanceManager implements HostedInstanceMan
         return this.doRun(pluginUri, port);
     }
 
-    async debug(pluginUri: URI, debugConfig: DebugConfiguration): Promise<URI> {
+    async debug(pluginUri: URI, debugConfig: DebugPluginConfiguration): Promise<URI> {
         return this.doRun(pluginUri, undefined, debugConfig);
     }
 
-    private async doRun(pluginUri: URI, port?: number, debugConfig?: DebugConfiguration): Promise<URI> {
+    private async doRun(pluginUri: URI, port?: number, debugConfig?: DebugPluginConfiguration): Promise<URI> {
         if (this.isPluginRunning) {
             this.hostedPluginSupport.sendLog({ data: 'Hosted plugin instance is already running.', type: LogType.Info });
             throw new Error('Hosted instance is already running.');
@@ -142,8 +144,7 @@ export abstract class AbstractHostedInstanceManager implements HostedInstanceMan
             throw new Error('Not supported plugin location: ' + pluginUri.toString());
         }
 
-        this.instanceUri = await this.postProcessInstanceUri(
-            await this.runHostedPluginTheiaInstance(command, processOptions));
+        this.instanceUri = await this.postProcessInstanceUri(await this.runHostedPluginTheiaInstance(command, processOptions));
         this.pluginUri = pluginUri;
         // disable redirect to grab the release
         this.instanceOptions = {
@@ -215,11 +216,7 @@ export abstract class AbstractHostedInstanceManager implements HostedInstanceMan
             const url = this.instanceUri.toString();
             request.head(url, this.instanceOptions).on('response', res => {
                 // Wait that the status is OK
-                if (res.statusCode === 200) {
-                    resolve(true);
-                } else {
-                    resolve(false);
-                }
+                resolve(res.statusCode === 200);
             }).on('error', error => {
                 resolve(false);
             });
@@ -240,7 +237,7 @@ export abstract class AbstractHostedInstanceManager implements HostedInstanceMan
         return false;
     }
 
-    protected async getStartCommand(port?: number, debugConfig?: DebugConfiguration): Promise<string[]> {
+    protected async getStartCommand(port?: number, debugConfig?: DebugPluginConfiguration): Promise<string[]> {
 
         const processArguments = process.argv;
         let command: string[];
@@ -267,19 +264,7 @@ export abstract class AbstractHostedInstanceManager implements HostedInstanceMan
         }
 
         if (debugConfig) {
-            let debugString = '--hosted-plugin-';
-            if (debugConfig.debugMode) {
-                debugString += debugConfig.debugMode;
-            } else {
-                debugString += 'inspect';
-            }
-
-            debugString += '=0.0.0.0';
-
-            if (debugConfig.port) {
-                debugString += ':' + debugConfig.port;
-            }
-            command.push(debugString);
+            command.push(`--hosted-plugin-${debugConfig.debugMode || 'inspect'}=0.0.0.0${debugConfig.debugPort ? ':' + debugConfig.debugPort : ''}`);
         }
         return command;
     }
@@ -373,17 +358,14 @@ export class NodeHostedPluginRunner extends AbstractHostedInstanceManager {
         return options;
     }
 
-    protected async getStartCommand(port?: number, config?: DebugConfiguration): Promise<string[]> {
+    protected async getStartCommand(port?: number, debugConfig?: DebugPluginConfiguration): Promise<string[]> {
         if (!port) {
-            if (process.env.HOSTED_PLUGIN_PORT) {
-                port = Number(process.env.HOSTED_PLUGIN_PORT);
-            } else {
-                port = 3030;
-            }
+            port = process.env.HOSTED_PLUGIN_PORT ?
+                Number(process.env.HOSTED_PLUGIN_PORT) :
+                (debugConfig?.debugPort ? Number(debugConfig.debugPort) : DEFAULT_HOSTED_PLUGIN_PORT);
         }
-        return super.getStartCommand(port, config);
+        return super.getStartCommand(port, debugConfig);
     }
-
 }
 
 @injectable()