+++ packages/plugin-ext/src/plugin/type-converters.ts
@@ -15,7 +15,7 @@
  ********************************************************************************/
 
 import * as theia from '@theia/plugin';
-import { Position as P, Range as R, SymbolInformation, SymbolKind as S } from '@theia/core/shared/vscode-languageserver-types';
+import * as lstypes from '@theia/core/shared/vscode-languageserver-types';
 import { URI } from './types-impl';
 import * as rpc from '../common/plugin-api-rpc';
 import {
@@ -108,10 +108,6 @@ export function fromSelection(selection: types.Selection): Selection {
 }
 
 export function toRange(range: model.Range): types.Range {
-    // if (!range) {
-    //     return undefined;
-    // }
-
     const { startLineNumber, startColumn, endLineNumber, endColumn } = range;
     return new types.Range(startLineNumber - 1, startColumn - 1, endLineNumber - 1, endColumn - 1);
 }
@@ -376,7 +372,12 @@ function convertTags(tags: types.DiagnosticTag[] | undefined): types.MarkerTag[]
     const markerTags: types.MarkerTag[] = [];
     for (const tag of tags) {
         switch (tag) {
-            case types.DiagnosticTag.Unnecessary: markerTags.push(types.MarkerTag.Unnecessary);
+            case types.DiagnosticTag.Unnecessary:
+                markerTags.push(types.MarkerTag.Unnecessary);
+                break;
+            case types.DiagnosticTag.Deprecated:
+                markerTags.push(types.MarkerTag.Deprecated);
+                break;
         }
     }
     return markerTags;
@@ -394,6 +395,16 @@ export function fromLocation(location: theia.Location): model.Location {
         uri: location.uri,
         range: fromRange(location.range)
     };
+}
+
+export function fromTextDocumentShowOptions(options: theia.TextDocumentShowOptions): model.TextDocumentShowOptions {
+    if (options.selection) {
+        return {
+            ...options,
+            selection: fromRange(options.selection),
+        };
+    }
+    return options as model.TextDocumentShowOptions;
 }
 
 export function fromDefinitionLink(definitionLink: theia.DefinitionLink): model.LocationLink {
@@ -635,7 +646,10 @@ export function isModelCallHierarchyItem(thing: any): thing is model.CallHierarc
     if (!thing) {
         return false;
     }
-    return false;
+    return isModelRange(thing.range)
+        && isModelRange(thing.selectionRange)
+        && isUriComponents(thing.uri)
+        && !!thing.name;
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -643,7 +657,8 @@ export function isModelCallHierarchyIncomingCall(thing: any): thing is model.Cal
     if (!thing) {
         return false;
     }
-    return false;
+    const maybeIncomingCall = thing as model.CallHierarchyIncomingCall;
+    return 'from' in maybeIncomingCall && 'fromRanges' in maybeIncomingCall && isModelCallHierarchyItem(maybeIncomingCall.from);
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -651,7 +666,8 @@ export function isModelCallHierarchyOutgoingCall(thing: any): thing is model.Cal
     if (!thing) {
         return false;
     }
-    return false;
+    const maybeOutgoingCall = thing as model.CallHierarchyOutgoingCall;
+    return 'to' in maybeOutgoingCall && 'fromRanges' in maybeOutgoingCall && isModelCallHierarchyItem(maybeOutgoingCall.to);
 }
 
 export function toLocation(value: model.Location): types.Location {
@@ -959,29 +975,30 @@ export function getShellExecutionOptions(options: theia.ShellExecutionOptions): 
     return result;
 }
 
-export function fromSymbolInformation(symbolInformation: theia.SymbolInformation): SymbolInformation | undefined {
+export function fromSymbolInformation(symbolInformation: theia.SymbolInformation): lstypes.SymbolInformation | undefined {
     if (!symbolInformation) {
         return undefined;
     }
 
     if (symbolInformation.location && symbolInformation.location.range) {
-        const p1 = P.create(symbolInformation.location.range.start.line, symbolInformation.location.range.start.character);
-        const p2 = P.create(symbolInformation.location.range.end.line, symbolInformation.location.range.end.character);
-        return SymbolInformation.create(symbolInformation.name, symbolInformation.kind++ as S, R.create(p1, p2),
+        const p1 = lstypes.Position.create(symbolInformation.location.range.start.line, symbolInformation.location.range.start.character);
+        const p2 = lstypes.Position.create(symbolInformation.location.range.end.line, symbolInformation.location.range.end.character);
+        return lstypes.SymbolInformation.create(symbolInformation.name, symbolInformation.kind++ as lstypes.SymbolKind, lstypes.Range.create(p1, p2),
             symbolInformation.location.uri.toString(), symbolInformation.containerName);
     }
 
-    return <SymbolInformation>{
+    return {
         name: symbolInformation.name,
         containerName: symbolInformation.containerName,
-        kind: symbolInformation.kind++ as S,
+        kind: symbolInformation.kind++ as lstypes.SymbolKind,
         location: {
-            uri: symbolInformation.location.uri.toString()
+            uri: symbolInformation.location.uri.toString(),
+            range: symbolInformation.location.range,
         }
     };
 }
 
-export function toSymbolInformation(symbolInformation: SymbolInformation): theia.SymbolInformation | undefined {
+export function toSymbolInformation(symbolInformation: lstypes.SymbolInformation): theia.SymbolInformation | undefined {
     if (!symbolInformation) {
         return undefined;
     }